requirements.txt
streamlit
pandas
plotly
httpx-oauth
python-dotenv
typing-extensions
pdfrw
reportlab
openpyxl
kaleido
boto3
streamlit-nested-layout
streamlit-modal
python-dotenv
mysql-connector-python-rf
